requirements.txt
torch==2.1.0
bitsandbytes==0.41.1
tqdm==4.66.1
open_clip_torch==2.23.0
accelerate==0.24.1
wandb==0.15.12
pandas==2.1.2